i64->isize, u64->usize 	modified:   src/adapt_strategy.rs 	modified:   src/cpu_potential.rs 	modified:   src/cpu_sampler.rs 	modified:   src/cpu_state.rs 	modified:   src/mass_matrix.rs 	modified:   src/nuts.rs 	modified:   src/stepsize.rs

+++ src/cpu_potential.rs
@@ -70,11 +70,11 @@ impl<T: Send + Debug + Clone, E: Debug + Send + std::error::Error> DivergenceInf
         self.energy_error.clone()
     }
 
-    fn end_idx_in_trajectory(&self) -> Option<i64> {
+    fn end_idx_in_trajectory(&self) -> Option<isize> {
         Some(self.end.as_ref()?.idx_in_trajectory)
     }
 
-    fn start_idx_in_trajectory(&self) -> Option<i64> {
+    fn start_idx_in_trajectory(&self) -> Option<isize> {
         Some(self.end.as_ref()?.idx_in_trajectory)
     }
